#74 adding author credits to credits scene - WIP
Added a text field for credits and populated it with all downloaded assets that I've verified that are in the game so far. There may be a few tile sets or backgrounds that are not accounted for yet.  Will review them soon.

Also added the author name to our main character's art asset folder for easy reference.