#65 #88 testing animation event to trigger attack sound
Functionally this is working.
Leave the animator on the same sub object as the sprite renderer unless you wish to have to manually update the sprite on every key frame of every animation.
Added a script to the sprite object containing the test function we want our animation event to call.
Added same function names to player controller with actual behaviour.
Animator's animation event calls the function at the sprite renderer's level which bubble's it up to the player controller's level.

Seems to be working but may wish to play with the timing/placement of the animation event and add other effects other than audio.