#81 added simple collider object for stairs and slopes
Could go a cleaner route and may if needed as the game grows.

For now just added an empty object with a box collider 2d and manually scaled it to the stairs in the castle forest.

Not worried about any special handling of player speed or sliding at this time although it might be interesting if the player would float on stairs, unless that's a later ability that they float in general. Or can in a certain mode.