#97 added statue prefabs with hideable script - WIP
Added a prefab for a hideable statue and a broken hideable statue.

May integrate them into one later and just swap the sprite from broken to whole for a
quest.

Running into issues with the hideable script or settings though
My settings appear the same as the hay prefab, though the sprite sorting layer I believe differs from hay for the barrel and wagon.

When placing these statues in the castle scene, I can hide behind them but the alpha fade doesn't occur.

When playing the same prefabs in the first scene they fade fine.
- fixed by deleting statue instance in the castle forest scene and re-adding them, and setting the broken statue sprite additional layers sorting layer to default (same as the barrel and wagon, but for some reason hay works with this set to hideables while others don't)