#66 updated knockback to use animation events
Refactored knockback so that entities apply the knockback to themselves

The player hits an enemy and will call the enemies knockback function, telling it how far to be knocked back for now.
Likewise the enemy hits the player (when they can attack) and will call the player's knockback function.

The owner of the rigidbody is responsible for moving its own rigidbody, and no other rigidbody.

Added animation events for the damage point of the players attack swing and the end of the animation (last frame)

Collision is now checked, and damage applied, at time of the players attack animation reaching the bottom of its swing.

May wish to make knockback only occur in certain cases but will try it as is for now.