Attack Visual Effect
Imported a slash asset (from the free visual effects we found earlier- same artstyle as the halo effect on enemies). Made the slash effect a prefab. Modified the player animations by reslicing them and bottom centering them. Modified character controller so slash effect plays when the audio plays. Deleted duplicate animation sheets (for some reason every reaper animation had a duplicated sheet). Updated render layer with a new layer "VisualEffects" and placed it below ground so that it will appear above all else. Made the slash utilize this VisualEffects layer so it shows the full effect above all other sprites. Known issue: the visual effect seems to work well, but the actual attack logic is buggy at times where it doesn't coincide with how the visual plays, i.e if you attack and change your players direction quickly, the attack hitbox will spawn to the changed direction as opposed to where the visual is indicating. I suppose this can be changed 2 ways, fix the visual to match the attack logic or fix the attack logic. I feel that since the visual already coincides nicely with the audio logic, it would be better to fix how the attack is actually registering.